+++ Siliconweb/Controllers/AccountController.cs
@@ -4,5 +4,11 @@ namespace Siliconweb.Controllers
 {
     public class AccountController : Controller
     {
+        //private readonly AccountService _accountService;
+
+       // public AccountController(AccountService accountService)
+       // {
+        //    _accountService = accountService;
+       // }
     }
 }
