+++ Siliconweb/Controllers/AuthController.cs
@@ -1,5 +1,5 @@
 using Microsoft.AspNetCore.Mvc;
-using Siliconweb.Models.Views;
+using Siliconweb.ViewModels;
 
 namespace Siliconweb.Controllers;
 
@@ -33,24 +33,6 @@ public class AuthController : Controller
         var viewModel = new SignInViewModel();
         return View(viewModel);
     }
-
-    [Route("/signin")]
-    [HttpPost]
-    public IActionResult SignIn(SignInViewModel viewModel)
-    {
-        if (!ModelState.IsValid)
-        {
-            return View(viewModel);
-        }
-
-        // var result = _authService.SignIn(viewModel.Form);
-        // if (result)
-            //return RedirectToAction("Account", "Index");
-
-        viewModel.ErrorMessage = "Incorrect email or password";
-        return View(viewModel);
-
-       
-    }
+    
 
 }
