+++ Siliconweb/Models/Views/SignUpModel.cs
@@ -1,5 +1,6 @@
 using Microsoft.AspNetCore.Mvc.ModelBinding.Validation;
 using Siliconweb.Helpers;
+using Siliconweb.ViewModels;
 using System.ComponentModel.DataAnnotations;
 
 namespace Siliconweb.Models.Views;
