+++ Siliconweb/ViewModels/SignUpViewModel.cs
@@ -1,4 +1,6 @@
-namespace Siliconweb.Models.Views;
+using Siliconweb.Models.Views;
+
+namespace Siliconweb.ViewModels;
 
 public class SignUpViewModel
 {
@@ -7,5 +9,5 @@ public class SignUpViewModel
 
     public bool TermsAndConditions { get; set; } = false;
 
-    
+
 }
