+++ Siliconweb/Views/Shared/Sections/_FooterPartial.cshtml
@@ -1,0 +1,15 @@
+﻿
+    <footer>
+        <div class="container">
+            <div class="social-media">
+                <a class="btn-social" href="#"><i class="fa-brands fa-facebook"></i></a>
+                <a class="btn-social" href="#"><i class="fa-brands fa-square-x-twitter"></i></a>
+                <a class="btn-social" href="#"><i class="fa-brands fa-instagram"></i></a>
+                <a class="btn-social" href="#"><i class="fa-brands fa-youtube"></i></a>
+            </div>
+            <p>
+                &copy; 2024 Silicon, All rights reserved. Pulvinar urna condimentum amet tempor, ornare integer. Lorem
+                bla bla bla
+            </p>
+        </div>
+    </footer>
