+++ Siliconweb/Views/Shared/Sections/_HeaderPartial.cshtml
@@ -1,0 +1,29 @@
+﻿@*
+    For more information on enabling MVC for empty projects, visit https://go.microsoft.com/fwlink/?LinkID=397860
+*@
+@{
+    <header>
+        <div class="container">
+            <a id="logo" href="home.html"><img src="~/images/silicon-logo-light_theme.svg" alt=""> </a>
+            <div id="menu">
+                <nav>
+                    <a class="menu-link" href="#showcase">Showcase</a>
+                    <a class="menu-link" href="#features">Features</a>
+                    <a class="menu-link" href="contact.html">Contact</a>
+                </nav>
+            </div>
+            <div class="btn-switch">
+                <label>Light</label>
+                <label class="switch" for="switch-mode">
+                    <input type="checkbox" id="switch-mode" />
+                    <div class="slider round"></div>
+                </label>
+            </div>
+            <label>Dark</label>
+            <div class="account-buttons">
+                <a class="btn-gray" asp-controller="Auth" asp-action="SignIn"><i class="fa-regular fa-right-to-bracket"></i><span>Sign in</span></a>
+                <a class="btn-theme" asp-controller="Auth" asp-action="SignUp"><i class="fa-regular fa-user"></i><span>Sign up</span></a>
+                </div>
+        </div>
+    </header>
+}
