+++ Siliconweb/Views/Shared/_Layout.cshtml
@@ -11,48 +11,14 @@
 </head>
 <body>
 
-<header>
-        <div class="container">
-            <a id="logo" href="home.html"><img src="~/images/silicon-logo-light_theme.svg" alt=""> </a>
-            <div id="menu">
-                <nav>
-                    <a class="menu-link" href="#showcase">Showcase</a>
-                    <a class="menu-link" href="#features">Features</a>
-                    <a class="menu-link" href="contact.html">Contact</a>
-                </nav>
-            </div>
-            <div class="btn-switch">
-                <label>Light</label>
-                <label class="switch" for="switch-mode">
-                    <input type="checkbox" id="switch-mode" />
-                    <div class="slider round"></div>
-                </label>
-            </div>
-            <label>Dark</label>
-            <div class="account-buttons">
-                <a class="btn-gray" asp-controller="Auth" asp-action="SignIn"><i class="fa-regular fa-right-to-bracket"></i><span>Sign in</span></a>
-                <a  class="btn-theme" asp-controller="Auth" asp-action="SignUp"><i class="fa-regular fa-user"></i><span>Sign up</span></a>
-            </div>
-        </div>
-</header>
+    <div class="wrapper">
+        @await Html.PartialAsync("/Views/Shared/Sections/_HeaderPartial.cshtml")
 <main>
     @RenderBody()
 </main>
 
-<footer>
-        <div class="container">
-            <div class="social-media">
-                <a class="btn-social" href="#"><i class="fa-brands fa-facebook"></i></a>
-                <a class="btn-social" href="#"><i class="fa-brands fa-square-x-twitter"></i></a>
-                <a class="btn-social" href="#"><i class="fa-brands fa-instagram"></i></a>
-                <a class="btn-social" href="#"><i class="fa-brands fa-youtube"></i></a>
-            </div>
-            <p>
-                &copy; 2024 Silicon, All rights reserved. Pulvinar urna condimentum amet tempor, ornare integer. Lorem
-                bla bla bla
-            </p>
-        </div>
-</footer>
+        @await Html.PartialAsync("/Views/Shared/Sections/_FooterPartial.cshtml")
+     </div>
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/js/bootstrap.min.js" integrity="sha384-0pUGZvbkm6XF6gxjEnlmuGrJXVbNuzT9qBBavbLwCsOGabYfZo0T0to5eqruptLy" crossorigin="anonymous"></script>
 
 </body>
