+++ Siliconweb/Views/_ViewImports.cshtml
@@ -1,6 +1,5 @@
 ﻿@using Siliconweb
 @using Siliconweb.Models.Sections
 @using Siliconweb.Models.Views
+@using Siliconweb.ViewModels;
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
-@using Siliconweb.Models.Sections
-@using Siliconweb.Models.Views
